requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
torch==2.7.1
transformers==4.55.2
peft==0.17.0
safetensors==0.4.5
pandas==2.1.3
requests==2.31.0
mecab-python3==1.0.6
python-multipart==0.0.6
sentencepiece==0.1.99
protobuf==4.25.1
sacremoses==0.0.53
